Fix service methods not having access to `self`
This is to do with us passing in the player as the first argument to the method, instead of the service's table.

We're now passing in the service table first, and completely ignoring the player. ROBLOX's services don't automatically take in the player, and ours don't need to either.

If we ever have need of a player in the service, we can manually pass it in.